Removed use case package and classes, moved logic to the service
