+++ README.md
@@ -75,4 +75,5 @@ Also, value-objects are responsible for a little more than just plain data holdi
 * Replace Spring Framework with a different one.
 * The proposed architecture is not perfect. Suggest improvements.
     * <s>Add service layer for the AccountController</s>
+    * <s>Remove use case package and classes. Move logic to the service</s>
     * <s>Move configuration to separate file</s>
