+++ src/main/java/pl/cleankod/exchange/config/ApplicationConfiguration.java
@@ -11,14 +11,12 @@ import org.springframework.core.env.Environment;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 import pl.cleankod.exchange.core.gateway.AccountRepository;
 import pl.cleankod.exchange.core.gateway.CurrencyConversionService;
-import pl.cleankod.exchange.core.mapper.AccountMapper;
-import pl.cleankod.exchange.core.mapper.MoneyMapper;
 import pl.cleankod.exchange.core.service.AccountService;
 import pl.cleankod.exchange.core.service.AccountServiceImpl;
+import pl.cleankod.exchange.core.service.mapper.AccountMapper;
+import pl.cleankod.exchange.core.service.mapper.MoneyMapper;
 import pl.cleankod.exchange.core.trace.ContextInitializeInterceptor;
 import pl.cleankod.exchange.core.trace.WebMvcConfig;
-import pl.cleankod.exchange.core.usecase.FindAccountAndConvertCurrencyUseCase;
-import pl.cleankod.exchange.core.usecase.FindAccountUseCase;
 import pl.cleankod.exchange.entrypoint.AccountController;
 import pl.cleankod.exchange.entrypoint.ExceptionHandlerAdvice;
 import pl.cleankod.exchange.provider.AccountInMemoryRepository;
@@ -47,21 +45,6 @@ public class ApplicationConfiguration {
     @Bean
     CurrencyConversionService currencyConversionService(ExchangeRatesNbpClient exchangeRatesNbpClient) {
         return new CurrencyConversionNbpService(exchangeRatesNbpClient);
-    }
-
-    @Bean
-    FindAccountUseCase findAccountUseCase(AccountRepository accountRepository) {
-        return new FindAccountUseCase(accountRepository);
-    }
-
-    @Bean
-    FindAccountAndConvertCurrencyUseCase findAccountAndConvertCurrencyUseCase(
-            AccountRepository accountRepository,
-            CurrencyConversionService currencyConversionService,
-            Environment environment
-    ) {
-        Currency baseCurrency = Currency.getInstance(environment.getRequiredProperty("app.base-currency"));
-        return new FindAccountAndConvertCurrencyUseCase(accountRepository, currencyConversionService, baseCurrency);
     }
 
     @Bean
@@ -95,9 +78,13 @@ public class ApplicationConfiguration {
     }
 
     @Bean
-    AccountService accountService(FindAccountAndConvertCurrencyUseCase findAccountAndConvertCurrencyUseCase,
-                                  FindAccountUseCase findAccountUseCase,
-                                  AccountMapper accountMapper) {
-        return new AccountServiceImpl(findAccountAndConvertCurrencyUseCase, findAccountUseCase, accountMapper);
+    AccountService accountService(AccountMapper accountMapper,
+                                  CurrencyConversionService currencyConversionService,
+                                  AccountRepository accountRepository,
+                                  Environment environment) {
+
+        Currency baseCurrency = Currency.getInstance(environment.getRequiredProperty("app.base-currency"));
+
+        return new AccountServiceImpl(accountMapper, currencyConversionService, accountRepository, baseCurrency);
     }
 }
