--- src/main/java/pl/cleankod/exchange/core/usecase/FindAccountAndConvertCurrencyUseCase.java
@@ -1,46 +1,0 @@
-package pl.cleankod.exchange.core.usecase;
-
-import pl.cleankod.exchange.core.domain.Account;
-import pl.cleankod.exchange.core.domain.Money;
-import pl.cleankod.exchange.core.gateway.AccountRepository;
-import pl.cleankod.exchange.core.gateway.CurrencyConversionService;
-
-import java.util.Currency;
-import java.util.Optional;
-
-public class FindAccountAndConvertCurrencyUseCase {
-
-    private final AccountRepository accountRepository;
-    private final CurrencyConversionService currencyConversionService;
-    private final Currency baseCurrency;
-
-    public FindAccountAndConvertCurrencyUseCase(AccountRepository accountRepository,
-                                                CurrencyConversionService currencyConversionService,
-                                                Currency baseCurrency) {
-        this.accountRepository = accountRepository;
-        this.currencyConversionService = currencyConversionService;
-        this.baseCurrency = baseCurrency;
-    }
-
-    public Optional<Account> execute(Account.Id id, Currency targetCurrency) {
-        return accountRepository.find(id)
-                .map(account -> new Account(account.id(), account.number(), convert(account.balance(), targetCurrency)));
-    }
-
-    public Optional<Account> execute(Account.Number number, Currency targetCurrency) {
-        return accountRepository.find(number)
-                .map(account -> new Account(account.id(), account.number(), convert(account.balance(), targetCurrency)));
-    }
-
-    private Money convert(Money money, Currency targetCurrency) {
-        if (!baseCurrency.equals(targetCurrency)) {
-            return money.convert(currencyConversionService, targetCurrency);
-        }
-
-        if (!money.currency().equals(targetCurrency)) {
-            throw new CurrencyConversionException(money.currency(), targetCurrency);
-        }
-
-        return money;
-    }
-}
